# Conflicts:
#	src/main/java/net/Y5M2/admin/biz/AdminBiz.java
#	src/main/java/net/Y5M2/admin/biz/AdminBizImpl.java
#	src/main/java/net/Y5M2/admin/web/ViewAdminArticlePageServlet.java
#	src/main/java/net/Y5M2/admin/web/ViewAdminArticleWritePageServlet.java
#	src/main/java/net/Y5M2/article/dao/BoardDao.java
#	src/main/java/net/Y5M2/article/dao/BoardDaoImpl.java
#	src/main/java/net/Y5M2/category/biz/CategoryBizImpl.java
#	src/main/java/net/Y5M2/category/dao/CategoryDaoImpl.java
#	src/main/java/net/Y5M2/user/dao/UserDaoImpl.java
#	src/main/webapp/WEB-INF/view/admin/adminArticle.jsp

+++ src/main/webapp/WEB-INF/view/team/detailTeamInfo.jsp
@@ -1,5 +1,7 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+	pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+
 
 <jsp:include page="/WEB-INF/view/commons/header.jsp"></jsp:include>
 
@@ -17,25 +19,60 @@
 			$("#passwordConfirm").click(function(){
 
 				
-			if($("#password").val()=="${userInfo.password}"){
-				window.open("/ShootBoy/teamModify","","width=800, height=700");
-
-			}
-			else{
-				$("div.warning").html("<p>비밀번호가 틀렸습니다.</p>")
-			}
-
+				$.post("/ShootBoy/doCheckPassword", {"password" : $("#password").val()} ,function(data){
+					if(data == "true"){
+						window.open("/ShootBoy/teamModify","","width=800, height=800")
+					}
+					else{
+						alert("비밀번호가 틀렸습니다.")
+					}
+				});
 			});
 		});
 		
 		$("#deleteTeam").click(function () {
-			if ( confirm("정말로 팀을 해체 하시겠습니까?") ) {
-				alert("정상적으로 처리 되었습니다.");
-				location.href="/ShootBoy/doDeleteTeam?teamId=" + "${teamInfo.teamId}";
-			}
-		})
-	});
+			$("#password").show();
+			$("#passwordConfirm").show();
+			
+			$("#passwordConfirm").click(function(){
+				$.post("/ShootBoy/doCheckPassword", {"password" : $("#password").val()} ,function(data){
+					if(data == "true"){
+						if ( confirm("정말로 팀을 해체 하시겠습니까?") ) {
+							alert("정상적으로 처리 되었습니다.");
+							location.href="/ShootBoy/doDeleteTeam?teamId=" + "${teamInfo.teamId}";
+						}
+					}
+					else{
+						alert("비밀번호가 틀렸습니다.")
+					}
+				});
+				
+			});
+
+		});
 		
+		$("#dropTeam").click(function () {
+			$("#password").show();
+			$("#passwordConfirm").show();
+			
+			$("#passwordConfirm").click(function(){
+				$.post("/ShootBoy/doCheckPassword", {"password" : $("#password").val()} ,function(data){
+					if(data == "true"){
+						if ( confirm("정말로 팀을 탈퇴 하시겠습니까?") ) {
+							alert("정상적으로 처리 되었습니다.");
+							location.href="/ShootBoy/doDrop?teamId=" + "${teamInfo.teamId}";
+						}
+					}
+					else{
+						alert("비밀번호가 틀렸습니다.")
+					}
+				});
+				
+			});
+		});
+		
+		
+	});	
 </script>
 	<div id="myPageLeftMenu">
 		<div class="mplmTitle">My Page</div>
@@ -49,13 +86,14 @@
 	
 	<div class="myInfoText" style="display: inline-block; width: 800px;"><h1>팀 정보</h1>
 		<hr class="myPageline">
-		<div id="teamDetailwrapper">
+		<div id="teamDetailwrapper" style="padding-top: 20px;">
 		<c:if test="${!empty userInfo.teamId}">
 			<div id="teamImg" style="display: inline-block;">
 				<img src="/ShootBoy/showImage?teamId=${userInfo.teamId}" 
-						style="width: 300px; height: 300px; padding-top: 20px;
+						style="width: 300px; height: 300px;
 						border-radius: 150px;">
-			</div>	
+			<div class="teamPhoto" style="font-size: 30px; width: 150px; margin-left: 80px; padding-top: 20px;">팀 앰블럼</div>	
+			</div>
 			<div id="teamDetailContainer" style="display: inline-block; width: 400px; 
 					vertical-align: top; margin-left: 50px; padding-top: 20px;">
 				<div class="teamName" style="font-size: 20px; padding-bottom: 10px">팀명 : ${userInfo.teamVO.teamName}</div>
@@ -67,12 +105,33 @@
 			</div>
 			<div id="teamModify" style=" float: right; margin-right: 100px;"><input type="button" id="modifyBtn" name="modifyBtn" value="수정" 
 					style="margin-top: 20px; width: 200px;" >
-			<div> <input type="button" id="deleteTeam" name="deleteTeam" value="팀 해체" /> </div>
+
+					<c:choose>
+						<c:when test="${userInfo.levelId eq '3'}">
+							<div> <input type="button" id="deleteTeam" name="deleteTeam" value="팀 해체" /> </div>
+						</c:when>
+						<c:otherwise>
+							<div> <input type="button" id="dropTeam" name="dropTeam" value="팀 탈퇴" /> </div>
+						</c:otherwise>
+					</c:choose>
+					
+
 			<div><input type="password" id="password" name="password" placeholder="비밀번호를 입력해주세요">
 				<input type="button" id="passwordConfirm" name="passwordConfirm" value="비밀번호 확인" style="margin-left: 30px;">
 			</div>
 			</div>
-		<div class="warning"></div>
+		</c:if>
+		<c:if test="${sessionScope._USER_INFO_.levelId eq 3}">
+			<c:forEach items="${joins}" var="join" >
+			<tr>
+				<td>${join.userVO.userName}</td>
+				<td>${join.userVO.email}</td>
+				<td>${join.userVO.phoneNumber}</td>
+				<td>${join.userVO.age}</td>
+				<td>${join.userVO.position}</td>
+				<td>${join.userVO.locationVO.locationName}</td>
+			</tr>
+			</c:forEach>
 		</c:if>
 		<c:if test="${empty userInfo.teamId}">
 		<p style="font-size: 20px; color: red;">팀에 가입하지 않았습니다!</p>
